Allow different template directory
A user chooses a template via '--template' the program then looks for
this file inside the current working directory + 'templates' +
templatename. In case the GONTRIB_TEMPLATES_PTH environment variable is
set it will look in GONTRIB_TEMPLATES_PTH + templatename.

exconf command now prints the example configuration instead of writing
it to a file. In this way a user still can pipe it into a file himself.