Have log file when scan fails
Have an error.log file which logs all the repos that couldnt be cloned
and wikis that we couldnt get a count from.
In case the output list is very long we might miss this in the overview
so it would be good to have this in a seperate file.

So far we only printed that there was an error but not which one:
-util.PrintInfo("Problem loading repo", util.PI_MILD_ERROR)
+ util.PrintInfoF(logwriter, "Problem loading repo %s: %s", util.PI_MILD_ERROR, repo, err.Error())

If scanWiki count is 0 for a username it could also be that we mistyped
the username in the config file, so treat this as a message worth
mentioning.

scanOBS obs.GetLatestRepo so far only returned an error but didnt print
anything if there was a problem cloning/updating an OBS repo. Fixed not
and behaves like the others.

ScanContributions when running the specific scans (like scanGit) dont exit upon error because we want it just to continue and write the error to the log.  So far all the errors that happen in the specific scans are ones that get outputted and are non critical, so this is okay.

Add notice about error.log, if it exists to program output upon exit.

Solves #15.