Make PrintInfoF testable
PrintInfo() took a *bufio.Writer. Let's make it an io.Writer since
bufio.Writer satisfies the io.Writer interface.
Meaning inside we have to use Fprintf() instead of WriteString().

Before, we did the testing whether we want to print to stdout or to the
logger inside the PrintInfo(), very ugly.

Now we pass os.Stdout to it or the logwriter or..